refactor investment chart data work
